Viz/requirements.txt
watchdog
google-generativeai
python-dotenv

